Refactor all routes to AppRouter
